QA-12764 refactor: add resolveEserviceId in AbstractResolver
