+++ interop-qa-tests/src/test/java/it/pagopa/pn/interop/cucumber/steps/m2m/client/utils/ClientResolver.java
@@ -3,14 +3,17 @@ package it.pagopa.pn.interop.cucumber.steps.m2m.client.utils;
 import it.pagopa.pn.interop.cucumber.steps.SharedStepsContext;
 import it.pagopa.pn.interop.cucumber.steps.m2m.common.utils.AbstractResolver;
 import it.pagopa.pn.interop.cucumber.utility.StepParser;
-import lombok.RequiredArgsConstructor;
 
 import java.util.UUID;
 
-@RequiredArgsConstructor
 public class ClientResolver extends AbstractResolver {
 
     private final SharedStepsContext sharedStepsContext;
+
+    public ClientResolver(SharedStepsContext sharedStepsContext) {
+        super(sharedStepsContext);
+        this.sharedStepsContext = sharedStepsContext;
+    }
 
     public UUID resolveClientId(String raw) {
         return resolveOrParse(
